Decimal value fix for POS and Domestic international service

+++ src/main/java/com/unisys/trans/cps/middleware/models/response/DomesticInternationalResponseDTO.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Data
 public class DomesticInternationalResponseDTO {
 
-    private Long totalValue;
+    private float totalValue;
 
     List<TopDomesticInternationalResponseDTO> topDomesticInternationalResponseDTOList;
